chore(constants): move `constants` to `app/app-config`
