+++ src/app/app-config.ts
@@ -1,5 +1,5 @@
 //=========================================================
-//  CONSTANTS
+//  APP CONFIG
 //---------------------------------------------------------
 export const APP_NAME = 'soundcloud-ngrx';
