+++ src/app/player/audio-service.ts
@@ -4,7 +4,7 @@ import 'rxjs/add/operator/map';
 
 import { Action } from '@ngrx/store';
 import { Observable } from 'rxjs/Observable';
-import { PLAYER_MAX_VOLUME, PLAYER_VOLUME_INCREMENT } from 'src/constants';
+import { PLAYER_MAX_VOLUME, PLAYER_VOLUME_INCREMENT } from 'src/app/app-config';
 import { Times } from './state/times-state';
 import { AudioSource } from './audio-source';
 import { PlayerActions } from './player-actions';
