+++ src/app/tracklists/state/selectors.ts
@@ -6,7 +6,7 @@ import 'rxjs/add/operator/withLatestFrom';
 
 import { List } from 'immutable';
 import { AppState } from 'src/app';
-import { TRACKS_PER_PAGE } from 'src/constants';
+import { TRACKS_PER_PAGE } from 'src/app/app-config';
 import { Selector } from 'src/app/core';
 import { Tracklist } from '../models/tracklist';
 import { Track } from '../models/track';
