+++ src/app/tracklists/state/tracklist-reducer.ts
@@ -1,6 +1,6 @@
 import { Action, ActionReducer } from '@ngrx/store';
 import { List } from 'immutable';
-import { TRACKS_PER_PAGE } from 'src/constants';
+import { TRACKS_PER_PAGE } from 'src/app/app-config';
 import { SearchActions } from 'src/app/search/search-actions';
 import { UserActions } from 'src/app/users/user-actions';
 import { TrackData } from '../models/track';
